Rework React configuration handling
It is the responsibility of Drupal to provide JSON configuration in the
format expected by the React components.
https://github.com/danskernesdigitalebibliotek/dpl-react/blob/develop/docs/architecture/adr-007-configuration.md

This is especially the case for Instant Loans which make use of JSON
formats for configuration.

To ensure this we clarify the responsibility of the loadConfig() and
getConfig methods for React config implementations.

This is then applied to the Instant Loan Settings in particular.

All other settings can use a legacy implementation. This should not
matter much since they do not provide config in JSON formats.